refactor: 빈설정 제외 및 util클래스로 분리
