--- src/main/java/org/sopt/carena/diet/application/port/in/EmbeddingTextGenerateUseCase.java
@@ -1,7 +1,0 @@
-package org.sopt.carena.diet.application.port.in;
-
-import org.sopt.carena.diet.domain.DietChunk;
-
-public interface EmbeddingTextGenerateUseCase {
-    String generate(DietChunk chunk,String documentTitle);
-}
